Aim assist is implemented
When the player "shoots" their gun, 100 hitscan lines (tracers) are drawn inside the cone that they can fire within. On the first tracer hit, the loop is broken and no more tracers are fired. The firing cone is then effectively rotated to be centered on the tracer that hit. This is done by adding the rotation from the cone firing mechanic to the rotation that was already applied by the line.